pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[project]
authors = [{ name = "Gergely Wootsch", email = "hello@gergely-wootsch.com" }]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.13",
  "Topic :: Software Development :: Documentation",
  "Topic :: Software Development :: Quality Assurance",
  "Typing :: Typed",
]
dependencies = [
  "PyYAML>=6.0.3",
  "httpx>=0.28.1",
  "mcp>=1.26.0",
  "pydantic>=2.12.5",
  "rich>=14.3.2",
  "sse-starlette>=3.2.0",
  "starlette>=0.52.1",
  "typer>=0.12.0",
  "uvicorn>=0.41.0",
]
description = "A governed development framework for AI-assisted engineering"
keywords = [
  "acp",
  "adr",
  "ai-assisted",
  "governance",
  "mcp",
  "spec-driven-development",
]
license = { file = "LICENSE" }
name = "vaultspec-core"
readme = "README.md"
requires-python = ">=3.13"
version = "0.1.0"

  [project.urls]
  "Bug Tracker" = "https://github.com/wgergely/vaultspec-core/issues"
  Documentation = "https://github.com/wgergely/vaultspec-core/tree/main/.vaultspec/docs"
  Homepage = "https://github.com/wgergely/vaultspec-core"
  Repository = "https://github.com/wgergely/vaultspec-core"

  [project.scripts]
  "vaultspec-core" = "vaultspec_core.__main__:main"
  "vaultspec-mcp" = "vaultspec_core.mcp_server.app:run"

  [project.optional-dependencies]
  dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-timeout>=2.4.0",
    "pytest-reportlog>=0.4.0",
    "pytest-durations>=1.0.0",
    "ruff>=0.15.2",
    "ty>=0.0.15",
    "prek>=0.3.2",
    "identify>=2.0.0",
    "nodeenv>=1.0.0",
  ]

[dependency-groups]
dev = [
  "identify>=2.0.0",
  "nodeenv>=1.0.0",
  "prek>=0.3.2",
  "pytest>=9.0.2",
  "pytest-asyncio>=1.3.0",
  "pytest-durations>=1.0.0",
  "pytest-reportlog>=0.4.0",
  "pytest-timeout>=2.4.0",
  "ruff>=0.15.2",
  "ty>=0.0.15",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/vaultspec_core"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
log_cli = true
log_cli_date_format = "%H:%M:%S"
log_cli_format = "%(asctime)s [%(levelname)8s] %(name)s: %(message)s"
log_cli_level = "INFO"
log_file = "test-debug.log"
log_file_level = "DEBUG"
markers = [
  "api: API-level tests covering stable public analysis surfaces",
  "unit: fast tests with no external dependencies",
  "integration: requires CLI tools and network",
  "gemini: requires Gemini CLI installed",
  "claude: requires Claude CLI installed",
  "timeout: timeout in seconds for individual tests",
  "e2e: end-to-end tests requiring real LLM API keys",
]
testpaths = ["tests", "src"]
timeout = 300
timeout_func_only = true

[tool.ruff]
line-length = 88
target-version = "py313"

  [tool.ruff.lint]
  select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "RUF", # ruff-specific rules
    "N",   # pep8-naming
  ]
  unfixable = ["B"]

    [tool.ruff.lint.per-file-ignores]
    "**/test_*.py" = ["ARG"]
    "**/tests/**/*.py" = ["ARG"]
    # Core CLI modules use argparse/Path as runtime parameter types — no need to move to TYPE_CHECKING
    "src/vaultspec_core/core/*.py" = ["TC003"]

    [tool.ruff.lint.isort]
    known-first-party = ["vaultspec_core"]

[tool.ty.environment]
python-version = "3.13"
root = [".", "src"]
